sui_indexer_alt_reader::consistent_reader

Module proto

Modules

consistent_service_client

Generated client implementations.

consistent_service_server

Generated server implementations.

owner

Nested message and enum types in Owner.

Structs

AvailableRangeRequest

AvailableRangeResponse

Balance

BatchGetBalancesRequest

BatchGetBalancesResponse

GetBalanceRequest

ListBalancesRequest

ListBalancesResponse

ListObjectsByTypeRequest

ListObjectsResponse

ListOwnedObjectsRequest

Object

Owner

Enum of different types of ownership for an object.

ServiceConfigRequest

ServiceConfigResponse

Enums

End

The side of the filtered range to fetch the next page from.

Constants

CHECKPOINT_METADATA

Metadata name used in requests to set the checkpoint to make the request at, and in responses to indicate the checkpoint at which the response was generated.

FILE_DESCRIPTOR_SET