sui_indexer_alt_reader

Module pg_reader

Re-exports

pub use sui_pg_db as db;

Structs

Connection

PgReader

This wrapper type exists to perform error conversion between the data fetching layer and the RPC layer, metrics collection, and debug logging of database queries.